Seção 9: DTOs e Data Annotation

+++ Back/src/BrnEventos.API/Controllers/EventosController.cs
@@ -1,9 +1,9 @@
-﻿using BrnEventos.Domain;
-using Microsoft.AspNetCore.Mvc;
+﻿using Microsoft.AspNetCore.Mvc;
 using BrnEventos.Application.Contratos;
 using System.Threading.Tasks;
 using System;
 using Microsoft.AspNetCore.Http;
+using BrnEventos.Application.Dtos;
 
 namespace BrnEventos.API.Controllers
 {
@@ -24,8 +24,7 @@ namespace BrnEventos.API.Controllers
             try
             {
                 var eventos = await _eventoService.GetAllEventosAsync(true);
-                if (eventos == null) return NotFound("Nenhum evento encontrado!");
-
+                if (eventos == null) return NoContent();
                 return Ok(eventos);
             }
             catch (Exception ex)
@@ -41,7 +40,7 @@ namespace BrnEventos.API.Controllers
             try
             {
                 var evento = await _eventoService.GetEventoByIdAsync(id, true);
-                if (evento == null) return NotFound("Evento por ID não encontrado!");
+                if (evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -58,7 +57,7 @@ namespace BrnEventos.API.Controllers
             try
             {
                 var evento = await _eventoService.GetAllEventosByTemaAsync(tema, true);
-                if (evento == null) return NotFound("Eventos por tema não encontrados!");
+                if (evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -70,12 +69,12 @@ namespace BrnEventos.API.Controllers
         }
 
         [HttpPost]
-        public async Task<IActionResult> Post(Evento model)
+        public async Task<IActionResult> Post(EventoDto model)
         {
             try
             {
                 var evento = await _eventoService.AddEventos(model);
-                if (evento == null) return BadRequest("Erro ao tentar adicionar Evento!");
+                if (evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -87,12 +86,12 @@ namespace BrnEventos.API.Controllers
         }
 
         [HttpPut("{id}")]
-        public async Task<IActionResult> Put(int id, Evento model)
+        public async Task<IActionResult> Put(int id, EventoDto model)
         {
             try
             {
                 var evento = await _eventoService.UpdateEvento(id, model);
-                if (evento == null) return BadRequest("Erro ao tentar atualizar evento!");
+                if (evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -108,9 +107,12 @@ namespace BrnEventos.API.Controllers
         {
             try
             {
+                var evento = await _eventoService.GetEventoByIdAsync(id, true);
+                if (evento == null) return NoContent();
+
                 return await _eventoService.DeleteEvento(id) ? 
                        Ok("Deletado") : 
-                       BadRequest("Evento não deletado");
+                       throw new Exception("Ocorreu um erro não específico ao tentar deletar o Evento");
             }
             catch (Exception ex)
             {
